♻️ Aside 떼어내기
